nopbai

+++ module2/src/_1_intruction_java/bai_tap/BaiTap3.java
@@ -6,8 +6,8 @@ public class BaiTap3 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Nhập giá trị Dollar: ");
-        float dollars = scanner.nextFloat();
-        float rate = dollars * 23000;
+        double dollars = scanner.nextDouble();
+        double rate = dollars * 23000.0;
         System.out.println("Giá chuyển đổi sang VNĐ là: " + rate + " VNĐ");
     }
 }
